feat(utils): publicly expose utils related to eslint-plugin
