+++ packages/eslint-plugin/src/rules/component-selector.ts
@@ -1,25 +1,12 @@
-import type { TSESTree } from '@typescript-eslint/experimental-utils';
-import { createESLintRule } from '../utils/create-eslint-rule';
-import type { Options } from '../utils/property-selector';
-import {
-  checkSelector,
-  checkValidOptions,
-  OPTION_TYPE_ATTRIBUTE,
-  OPTION_TYPE_ELEMENT,
-  reportPrefixError,
-  reportStyleError,
-  reportTypeError,
-} from '../utils/property-selector';
-import { COMPONENT_CLASS_DECORATOR } from '../utils/selectors';
-import type { SelectorStyle } from '../utils/utils';
 import {
   arrayify,
-  getDecoratorPropertyValue,
-  isMemberExpression,
-  OPTION_STYLE_CAMEL_CASE,
-  OPTION_STYLE_KEBAB_CASE,
-} from '../utils/utils';
-import { ASTUtils } from '@typescript-eslint/experimental-utils';
+  ASTUtils,
+  Selectors,
+  SelectorUtils,
+} from '@angular-eslint/utils';
+import type { TSESTree } from '@typescript-eslint/experimental-utils';
+import { ASTUtils as TSESLintASTUtils } from '@typescript-eslint/experimental-utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 
 const VIEW_ENCAPSULATION_SHADOW_DOM = 'ShadowDom';
 const VIEW_ENCAPSULATION = 'ViewEncapsulation';
@@ -37,7 +24,7 @@ const STYLE_GUIDE_TYPE_LINK = 'https://angular.io/guide/styleguide#style-05-03';
 const SHADOW_DOM_ENCAPSULATED_STYLE_LINK =
   'https://github.com/angular-eslint/angular-eslint/issues/534';
 
-export default createESLintRule<Options, MessageIds>({
+export default createESLintRule<SelectorUtils.Options, MessageIds>({
   name: RULE_NAME,
   meta: {
     type: 'suggestion',
@@ -57,7 +44,10 @@ export default createESLintRule<Options, MessageIds>({
               {
                 type: 'array',
                 items: {
-                  enum: [OPTION_TYPE_ELEMENT, OPTION_TYPE_ATTRIBUTE],
+                  enum: [
+                    SelectorUtils.OPTION_TYPE_ELEMENT,
+                    SelectorUtils.OPTION_TYPE_ATTRIBUTE,
+                  ],
                 },
               },
             ],
@@ -67,7 +57,10 @@ export default createESLintRule<Options, MessageIds>({
           },
           style: {
             type: 'string',
-            enum: [OPTION_STYLE_CAMEL_CASE, OPTION_STYLE_KEBAB_CASE],
+            enum: [
+              ASTUtils.OPTION_STYLE_CAMEL_CASE,
+              ASTUtils.OPTION_STYLE_KEBAB_CASE,
+            ],
           },
         },
         additionalProperties: false,
@@ -77,7 +70,7 @@ export default createESLintRule<Options, MessageIds>({
       prefixFailure: `The selector should start with one of these prefixes: {{prefix}} (${STYLE_GUIDE_PREFIX_LINK})`,
       styleFailure: `The selector should be {{style}} (${STYLE_GUIDE_STYLE_LINK})`,
       typeFailure: `The selector should be used as an {{type}} (${STYLE_GUIDE_TYPE_LINK})`,
-      shadowDomEncapsulatedStyleFailure: `The selector of a ShadowDom-encapsulated component should be \`${OPTION_STYLE_KEBAB_CASE}\` (${SHADOW_DOM_ENCAPSULATED_STYLE_LINK})`,
+      shadowDomEncapsulatedStyleFailure: `The selector of a ShadowDom-encapsulated component should be \`${ASTUtils.OPTION_STYLE_KEBAB_CASE}\` (${SHADOW_DOM_ENCAPSULATED_STYLE_LINK})`,
     },
   },
   defaultOptions: [
@@ -89,14 +82,21 @@ export default createESLintRule<Options, MessageIds>({
   ],
   create(context, [{ type, prefix, style }]) {
     return {
-      [COMPONENT_CLASS_DECORATOR](node: TSESTree.Decorator) {
-        const rawSelectors = getDecoratorPropertyValue(node, 'selector');
+      [Selectors.COMPONENT_CLASS_DECORATOR](node: TSESTree.Decorator) {
+        const rawSelectors = ASTUtils.getDecoratorPropertyValue(
+          node,
+          'selector',
+        );
 
         if (!rawSelectors) {
           return;
         }
 
-        const isValidOptions = checkValidOptions(type, prefix, style);
+        const isValidOptions = SelectorUtils.checkValidOptions(
+          type,
+          prefix,
+          style,
+        );
 
         if (!isValidOptions) {
           return;
@@ -104,16 +104,16 @@ export default createESLintRule<Options, MessageIds>({
 
         // override `style` for ShadowDom-encapsulated components. See https://github.com/angular-eslint/angular-eslint/issues/534.
         const overrideStyle =
-          style !== OPTION_STYLE_KEBAB_CASE &&
+          style !== ASTUtils.OPTION_STYLE_KEBAB_CASE &&
           hasEncapsulationShadowDomProperty(node)
-            ? OPTION_STYLE_KEBAB_CASE
+            ? ASTUtils.OPTION_STYLE_KEBAB_CASE
             : style;
 
-        const hasExpectedSelector = checkSelector(
+        const hasExpectedSelector = SelectorUtils.checkSelector(
           rawSelectors,
           type,
           arrayify<string>(prefix),
-          overrideStyle as SelectorStyle,
+          overrideStyle as ASTUtils.SelectorStyle,
         );
 
         if (hasExpectedSelector === null) {
@@ -121,10 +121,10 @@ export default createESLintRule<Options, MessageIds>({
         }
 
         if (!hasExpectedSelector.hasExpectedType) {
-          reportTypeError(rawSelectors, type, context);
+          SelectorUtils.reportTypeError(rawSelectors, type, context);
         } else if (!hasExpectedSelector.hasExpectedStyle) {
           if (style === overrideStyle) {
-            reportStyleError(rawSelectors, style, context);
+            SelectorUtils.reportStyleError(rawSelectors, style, context);
           } else {
             context.report({
               node: rawSelectors,
@@ -132,7 +132,7 @@ export default createESLintRule<Options, MessageIds>({
             });
           }
         } else if (!hasExpectedSelector.hasExpectedPrefix) {
-          reportPrefixError(rawSelectors, prefix, context);
+          SelectorUtils.reportPrefixError(rawSelectors, prefix, context);
         }
       },
     };
@@ -140,13 +140,16 @@ export default createESLintRule<Options, MessageIds>({
 });
 
 function hasEncapsulationShadowDomProperty(node: TSESTree.Decorator) {
-  const encapsulationValue = getDecoratorPropertyValue(node, 'encapsulation');
+  const encapsulationValue = ASTUtils.getDecoratorPropertyValue(
+    node,
+    'encapsulation',
+  );
   return (
     encapsulationValue &&
-    isMemberExpression(encapsulationValue) &&
-    ASTUtils.isIdentifier(encapsulationValue.object) &&
+    ASTUtils.isMemberExpression(encapsulationValue) &&
+    TSESLintASTUtils.isIdentifier(encapsulationValue.object) &&
     encapsulationValue.object.name === VIEW_ENCAPSULATION &&
-    ASTUtils.isIdentifier(encapsulationValue.property) &&
+    TSESLintASTUtils.isIdentifier(encapsulationValue.property) &&
     encapsulationValue.property.name === VIEW_ENCAPSULATION_SHADOW_DOM
   );
 }
