+++ packages/eslint-plugin/src/rules/directive-class-suffix.ts
@@ -1,12 +1,10 @@
+import {
+  ASTUtils,
+  Selectors,
+  toHumanReadableText,
+} from '@angular-eslint/utils';
 import type { TSESTree } from '@typescript-eslint/experimental-utils';
 import { createESLintRule } from '../utils/create-eslint-rule';
-import { DIRECTIVE_CLASS_DECORATOR } from '../utils/selectors';
-import {
-  getClassName,
-  getDeclaredInterfaceNames,
-  getDecoratorPropertyValue,
-  toHumanReadableText,
-} from '../utils/utils';
 
 type Options = [{ readonly suffixes: readonly string[] }];
 export type MessageIds = 'directiveClassSuffix';
@@ -45,8 +43,8 @@ export default createESLintRule<Options, MessageIds>({
   defaultOptions: [{ suffixes: DEFAULT_SUFFIXES }],
   create(context, [{ suffixes }]) {
     return {
-      [DIRECTIVE_CLASS_DECORATOR](node: TSESTree.Decorator) {
-        const selectorPropertyValue = getDecoratorPropertyValue(
+      [Selectors.DIRECTIVE_CLASS_DECORATOR](node: TSESTree.Decorator) {
+        const selectorPropertyValue = ASTUtils.getDecoratorPropertyValue(
           node,
           'selector',
         );
@@ -54,8 +52,9 @@ export default createESLintRule<Options, MessageIds>({
         if (!selectorPropertyValue) return;
 
         const classParent = node.parent as TSESTree.ClassDeclaration;
-        const className = getClassName(classParent);
-        const declaredInterfaceNames = getDeclaredInterfaceNames(classParent);
+        const className = ASTUtils.getClassName(classParent);
+        const declaredInterfaceNames =
+          ASTUtils.getDeclaredInterfaceNames(classParent);
         const hasValidatorInterface = declaredInterfaceNames.some(
           (interfaceName) => interfaceName.endsWith(VALIDATOR_SUFFIX),
         );
