+++ packages/eslint-plugin/src/rules/directive-selector.ts
@@ -1,23 +1,11 @@
-import type { TSESTree } from '@typescript-eslint/experimental-utils';
-import { createESLintRule } from '../utils/create-eslint-rule';
-import type { Options } from '../utils/property-selector';
-import {
-  checkSelector,
-  checkValidOptions,
-  OPTION_TYPE_ATTRIBUTE,
-  OPTION_TYPE_ELEMENT,
-  reportPrefixError,
-  reportStyleError,
-  reportTypeError,
-} from '../utils/property-selector';
-import { DIRECTIVE_CLASS_DECORATOR } from '../utils/selectors';
-import type { SelectorStyle } from '../utils/utils';
 import {
   arrayify,
-  getDecoratorPropertyValue,
-  OPTION_STYLE_CAMEL_CASE,
-  OPTION_STYLE_KEBAB_CASE,
-} from '../utils/utils';
+  ASTUtils,
+  Selectors,
+  SelectorUtils,
+} from '@angular-eslint/utils';
+import type { TSESTree } from '@typescript-eslint/experimental-utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 
 export const RULE_NAME = 'directive-selector';
 export type MessageIds = 'prefixFailure' | 'styleFailure' | 'typeFailure';
@@ -26,7 +14,7 @@ const STYLE_GUIDE_PREFIX_LINK =
 const STYLE_GUIDE_STYLE_TYPE_LINK =
   'https://angular.io/guide/styleguide#style-02-06';
 
-export default createESLintRule<Options, MessageIds>({
+export default createESLintRule<SelectorUtils.Options, MessageIds>({
   name: RULE_NAME,
   meta: {
     type: 'suggestion',
@@ -45,7 +33,10 @@ export default createESLintRule<Options, MessageIds>({
               {
                 type: 'array',
                 items: {
-                  enum: [OPTION_TYPE_ELEMENT, OPTION_TYPE_ATTRIBUTE],
+                  enum: [
+                    SelectorUtils.OPTION_TYPE_ELEMENT,
+                    SelectorUtils.OPTION_TYPE_ATTRIBUTE,
+                  ],
                 },
               },
             ],
@@ -55,7 +46,10 @@ export default createESLintRule<Options, MessageIds>({
           },
           style: {
             type: 'string',
-            enum: [OPTION_STYLE_CAMEL_CASE, OPTION_STYLE_KEBAB_CASE],
+            enum: [
+              ASTUtils.OPTION_STYLE_CAMEL_CASE,
+              ASTUtils.OPTION_STYLE_KEBAB_CASE,
+            ],
           },
         },
         additionalProperties: false,
@@ -76,24 +70,31 @@ export default createESLintRule<Options, MessageIds>({
   ],
   create(context, [{ type, prefix, style }]) {
     return {
-      [DIRECTIVE_CLASS_DECORATOR](node: TSESTree.Decorator) {
-        const rawSelectors = getDecoratorPropertyValue(node, 'selector');
+      [Selectors.DIRECTIVE_CLASS_DECORATOR](node: TSESTree.Decorator) {
+        const rawSelectors = ASTUtils.getDecoratorPropertyValue(
+          node,
+          'selector',
+        );
 
         if (!rawSelectors) {
           return;
         }
 
-        const isValidOptions = checkValidOptions(type, prefix, style);
+        const isValidOptions = SelectorUtils.checkValidOptions(
+          type,
+          prefix,
+          style,
+        );
 
         if (!isValidOptions) {
           return;
         }
 
-        const hasExpectedSelector = checkSelector(
+        const hasExpectedSelector = SelectorUtils.checkSelector(
           rawSelectors,
           type,
           arrayify<string>(prefix),
-          style as SelectorStyle,
+          style as ASTUtils.SelectorStyle,
         );
 
         if (hasExpectedSelector === null) {
@@ -101,11 +102,11 @@ export default createESLintRule<Options, MessageIds>({
         }
 
         if (!hasExpectedSelector.hasExpectedType) {
-          reportTypeError(rawSelectors, type, context);
+          SelectorUtils.reportTypeError(rawSelectors, type, context);
         } else if (!hasExpectedSelector.hasExpectedStyle) {
-          reportStyleError(rawSelectors, style, context);
+          SelectorUtils.reportStyleError(rawSelectors, style, context);
         } else if (!hasExpectedSelector.hasExpectedPrefix) {
-          reportPrefixError(rawSelectors, prefix, context);
+          SelectorUtils.reportPrefixError(rawSelectors, prefix, context);
         }
       },
     };
