+++ packages/eslint-plugin/src/rules/no-input-rename.ts
@@ -1,21 +1,14 @@
-import type { TSESTree } from '@typescript-eslint/experimental-utils';
-import { ASTUtils } from '@typescript-eslint/experimental-utils';
-import { createESLintRule } from '../utils/create-eslint-rule';
-import { getAriaAttributeKeys } from '../utils/get-aria-attribute-keys';
 import {
-  COMPONENT_OR_DIRECTIVE_SELECTOR_LITERAL,
-  INPUTS_METADATA_PROPERTY_LITERAL,
-  INPUT_ALIAS,
-} from '../utils/selectors';
-import {
+  ASTUtils,
   capitalize,
-  getNearestNodeFrom,
-  getRawText,
-  getReplacementText,
-  isClassPropertyOrMethodDefinition,
+  getAriaAttributeKeys,
   kebabToCamelCase,
+  Selectors,
   withoutBracketsAndWhitespaces,
-} from '../utils/utils';
+} from '@angular-eslint/utils';
+import type { TSESTree } from '@typescript-eslint/experimental-utils';
+import { ASTUtils as TSESLintASTUtils } from '@typescript-eslint/experimental-utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 
 type Options = [{ readonly allowedNames?: readonly string[] }];
 export type MessageIds =
@@ -65,28 +58,32 @@ export default createESLintRule<Options, MessageIds>({
     const ariaAttributeKeys = getAriaAttributeKeys();
 
     return {
-      [COMPONENT_OR_DIRECTIVE_SELECTOR_LITERAL](
+      [Selectors.COMPONENT_OR_DIRECTIVE_SELECTOR_LITERAL](
         node: TSESTree.Literal | TSESTree.TemplateElement,
       ) {
         selectors = new Set(
-          withoutBracketsAndWhitespaces(getRawText(node)).split(','),
+          withoutBracketsAndWhitespaces(ASTUtils.getRawText(node)).split(','),
         );
       },
-      [INPUT_ALIAS](node: TSESTree.Literal | TSESTree.TemplateElement) {
-        const classPropertyOrMethodDefinition = getNearestNodeFrom(
+      [Selectors.INPUT_ALIAS](
+        node: TSESTree.Literal | TSESTree.TemplateElement,
+      ) {
+        const classPropertyOrMethodDefinition = ASTUtils.getNearestNodeFrom(
           node,
-          isClassPropertyOrMethodDefinition,
+          ASTUtils.isClassPropertyOrMethodDefinition,
         );
 
         if (
           !classPropertyOrMethodDefinition ||
-          !ASTUtils.isIdentifier(classPropertyOrMethodDefinition.key)
+          !TSESLintASTUtils.isIdentifier(classPropertyOrMethodDefinition.key)
         ) {
           return;
         }
 
-        const aliasName = getRawText(node);
-        const propertyName = getRawText(classPropertyOrMethodDefinition.key);
+        const aliasName = ASTUtils.getRawText(node);
+        const propertyName = ASTUtils.getRawText(
+          classPropertyOrMethodDefinition.key,
+        );
 
         if (
           allowedNames.includes(aliasName) ||
@@ -125,11 +122,11 @@ export default createESLintRule<Options, MessageIds>({
           });
         }
       },
-      [INPUTS_METADATA_PROPERTY_LITERAL](
+      [Selectors.INPUTS_METADATA_PROPERTY_LITERAL](
         node: TSESTree.Literal | TSESTree.TemplateElement,
       ) {
         const [propertyName, aliasName] = withoutBracketsAndWhitespaces(
-          getRawText(node),
+          ASTUtils.getRawText(node),
         ).split(':');
 
         if (
@@ -146,7 +143,10 @@ export default createESLintRule<Options, MessageIds>({
             node,
             messageId: 'noInputRename',
             fix: (fixer) =>
-              fixer.replaceText(node, getReplacementText(node, propertyName)),
+              fixer.replaceText(
+                node,
+                ASTUtils.getReplacementText(node, propertyName),
+              ),
           });
         } else if (!isAliasNameAllowed(selectors, propertyName, aliasName)) {
           context.report({
@@ -160,7 +160,10 @@ export default createESLintRule<Options, MessageIds>({
             ).map(([messageId, name]) => ({
               messageId,
               fix: (fixer) =>
-                fixer.replaceText(node, getReplacementText(node, name)),
+                fixer.replaceText(
+                  node,
+                  ASTUtils.getReplacementText(node, name),
+                ),
             })),
           });
         }
